Trait sr_io::Externalities

pub trait Externalities<H> where
    H: Hasher, {
    fn storage(&self, key: &[u8]) -> Option<Vec<u8>>;

    fn original_storage(&self, key: &[u8]) -> Option<Vec<u8>>;

    fn child_storage(
        &self, 
        storage_key: ChildStorageKey<H>, 
        key: &[u8]
    ) -> Option<Vec<u8>>;

    fn kill_child_storage(&mut self, storage_key: ChildStorageKey<H>);

    fn clear_prefix(&mut self, prefix: &[u8]);

    fn place_storage(&mut self, key: Vec<u8>, value: Option<Vec<u8>>);

    fn place_child_storage(
        &mut self, 
        storage_key: ChildStorageKey<H>, 
        key: Vec<u8>, 
        value: Option<Vec<u8>>
    );

    fn chain_id(&self) -> u64;

    fn storage_root(&mut self) -> <H as Hasher>::Out
    where
        <H as Hasher>::Out: Ord;

    fn child_storage_root(&mut self, storage_key: ChildStorageKey<H>) -> Vec<u8>;

    fn storage_changes_root(
        &mut self, 
        parent: <H as Hasher>::Out, 
        parent_num: u64
    ) -> Option<<H as Hasher>::Out>
    where
        <H as Hasher>::Out: Ord;

    fn submit_extrinsic(&mut self, extrinsic: Vec<u8>) -> Result<(), ()>;

    fn storage_hash(&self, key: &[u8]) -> Option<<H as Hasher>::Out> { ... }

    fn original_storage_hash(&self, key: &[u8]) -> Option<<H as Hasher>::Out> { ... }

    fn set_storage(&mut self, key: Vec<u8>, value: Vec<u8>) { ... }

    fn set_child_storage(
        &mut self, 
        storage_key: ChildStorageKey<H>, 
        key: Vec<u8>, 
        value: Vec<u8>
    ) { ... }

    fn clear_storage(&mut self, key: &[u8]) { ... }

    fn clear_child_storage(
        &mut self, 
        storage_key: ChildStorageKey<H>, 
        key: &[u8]
    ) { ... }

    fn exists_storage(&self, key: &[u8]) -> bool { ... }

    fn exists_child_storage(
        &self, 
        storage_key: ChildStorageKey<H>, 
        key: &[u8]
    ) -> bool { ... }
}

Externalities: pinned to specific active address.

Required methods

fn storage(&self, key: &[u8]) -> Option<Vec<u8>>

Read runtime storage.

fn original_storage(&self, key: &[u8]) -> Option<Vec<u8>>

Read original runtime storage, ignoring any overlayed changes.

fn child_storage(
&self,
storage_key: ChildStorageKey<H>,
key: &[u8]
) -> Option<Vec<u8>>

Read child runtime storage.

fn kill_child_storage(&mut self, storage_key: ChildStorageKey<H>)

Clear an entire child storage.

fn clear_prefix(&mut self, prefix: &[u8])

Clear storage entries which keys are start with the given prefix.

fn place_storage(&mut self, key: Vec<u8>, value: Option<Vec<u8>>)

Set or clear a storage entry (key) of current contract being called (effective immediately).

fn place_child_storage(
&mut self,
storage_key: ChildStorageKey<H>,
key: Vec<u8>,
value: Option<Vec<u8>>
)

Set or clear a child storage entry. Return whether the operation succeeds.

fn chain_id(&self) -> u64

Get the identity of the chain.

fn storage_root(&mut self) -> <H as Hasher>::Out where
<H as Hasher>::Out: Ord,

Get the trie root of the current storage map. This will also update all child storage keys in the top-level storage map.

fn child_storage_root(&mut self, storage_key: ChildStorageKey<H>) -> Vec<u8>

Get the trie root of a child storage map. This will also update the value of the child storage keys in the top-level storage map. If the storage root equals the default hash as defined by the trie, the key in the top-level storage map will be removed.

fn storage_changes_root(
&mut self,
parent: <H as Hasher>::Out,
parent_num: u64
) -> Option<<H as Hasher>::Out> where
<H as Hasher>::Out: Ord,

Get the change trie root of the current storage overlay at a block with given parent.

fn submit_extrinsic(&mut self, extrinsic: Vec<u8>) -> Result<(), ()>

Submit extrinsic.

Returns an error in case the API is not available.

Loading content...
Provided methods

fn storage_hash(&self, key: &[u8]) -> Option<<H as Hasher>::Out>

Get storage value hash. This may be optimized for large values.

fn original_storage_hash(&self, key: &[u8]) -> Option<<H as Hasher>::Out>

Get original storage value hash, ignoring any overlayed changes. This may be optimized for large values.

fn set_storage(&mut self, key: Vec<u8>, value: Vec<u8>)

Set storage entry key of current contract being called (effective immediately).

fn set_child_storage(
&mut self,
storage_key: ChildStorageKey<H>,
key: Vec<u8>,
value: Vec<u8>
)

Set child storage entry key of current contract being called (effective immediately).

fn clear_storage(&mut self, key: &[u8])

Clear a storage entry (key) of current contract being called (effective immediately).

fn clear_child_storage(&mut self, storage_key: ChildStorageKey<H>, key: &[u8])

Clear a child storage entry (key) of current contract being called (effective immediately).

fn exists_storage(&self, key: &[u8]) -> bool

Whether a storage entry exists.

fn exists_child_storage(
&self,
storage_key: ChildStorageKey<H>,
key: &[u8]
) -> bool

Whether a child storage entry exists.

Loading content...